AI-service/index_serivce/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
banks==2.1.3
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.116.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
griffe==1.7.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
llama-index-cli==0.4.4
llama-index-core==0.12.48
llama-index-embeddings-ollama==0.6.0
llama-index-embeddings-openai==0.3.1
llama-index-graph-stores-neo4j==0.4.6
llama-index-instrumentation==0.2.0
llama-index-llms-ollama==0.6.2
llama-index-llms-openai==0.4.7
llama-index-vector-stores-qdrant==0.6.1
llama-index-workflows==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
minio==7.2.15
multidict==6.6.3
mypy_extensions==1.1.0
neo4j==5.28.1
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
ollama==0.5.1
openai==1.95.1
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
portalocker==2.10.1
propcache==0.3.2
protobuf==6.31.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.3
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wrapt==1.17.2
yarl==1.20.1
